+++ index.ts
@@ -1,2 +1,0 @@
-export { Latitude } from './nodes/Latitude/Latitude.node';
-export { LatitudeApi } from './credentials/LatitudeApi.credentials';